+++ src/store/index.js
@@ -25,6 +25,7 @@ export default new Vuex.Store({
     noaaState: false,
     currentZoomState: 4,
     selectedTimePeriodState: "&period=P7D",
+    showAllDisabled: true,
   },
   mutations: {
     toggleDrawerState(state, data) {
@@ -81,6 +82,9 @@ export default new Vuex.Store({
     getSelectedTimePeriodState(state, currentTimePeriod) {
       state.selectedTimePeriodState = currentTimePeriod;
     },
+    getShowAllDisabled(state, showAllDisabled) {
+      state.showAllDisabled = showAllDisabled;
+    },
   },
   getters: {
     drawerState: (state) => state.drawerState,
@@ -93,6 +97,7 @@ export default new Vuex.Store({
     noaaState: (state) => state.noaaState,
     currentZoomState: (state) => state.currentZoomState,
     selectedTimePeriodState: (state) => state.selectedTimePeriodState,
+    showAllDisabled: (state) => state.showAllDisabled,
   },
   actions: {},
   modules: {},